requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-ckeditor==6.0.0
django-debug-toolbar==2.2
django-jazzmin==2.2.8
django-js-asset==1.2.2
django-nested-admin==3.3.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.15
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Pillow==7.2.0
psycopg2-binary==2.8.6
python-decouple==3.3
python-monkey-business==1.0.0
pytz==2020.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
